Burst views/api/__init__.py into submodules #11894
